Chat example program
This is an example of what this would look like done as a compiled program - peer-to-peer chat app for realtime communication.  I wrote this years ago.  Btw, the password for this is adminpw.  Run this on 2 different computers and set them to connect to each other's IP or computer name.

It's this that I believe could be re-implemented using straight HTML/JavaScript/NodeJS.

Sorry, I don't have the source for this.  As I said, I wrote it years ago.